make env file
