+++ frontend/src/commands/envSample.ts
@@ -1,0 +1,1 @@
+export const apiEndPoint = '';
